Change `databases` package name to `connect`
